Move spaces wizard into plugin registry
