+++ apps/admin/src/modules/spaces/composables/index.ts
@@ -7,7 +7,6 @@ export * from './useSpaceEditForm';
 export * from './useSpaces';
 export * from './useSpacesActions';
 export * from './useSpacesDataSource';
-export * from './useSpacesOnboarding';
 export * from './useSpacesPlugin';
 export * from './useSpacesPlugins';
 export * from './useSpacesRefreshSignals';
