+++ apps/admin/src/modules/spaces/composables/types.ts
@@ -80,6 +80,7 @@ export interface IUseSpaceEditForm<TForm extends ISpaceEditForm = ISpaceEditForm
 export interface IUseSpacesPlugins {
 	plugins: ComputedRef<IPlugin<ISpacePluginsComponents, ISpacePluginsSchemas, ISpacePluginRoutes>[]>;
 	options: ComputedRef<{ value: IPluginElement['type']; label: string; disabled: boolean }[]>;
+	wizardOptions: ComputedRef<{ value: IPlugin['type']; label: string; description: string; disabled: boolean }[]>;
 	getByPluginType: (type: IPlugin['type']) => IPlugin<ISpacePluginsComponents, ISpacePluginsSchemas, ISpacePluginRoutes> | undefined;
 	getByType: (type: IPluginElement['type']) => IPlugin<ISpacePluginsComponents, ISpacePluginsSchemas, ISpacePluginRoutes> | undefined;
 	getElement: (type: IPluginElement['type']) => IPluginElement<ISpacePluginsComponents, ISpacePluginsSchemas> | undefined;
