+++ apps/admin/src/modules/spaces/views/view-spaces.vue
@@ -49,6 +49,7 @@
 		<template #extra>
 			<div class="flex items-center">
 				<el-button
+					v-if="wizardOptions.length > 0"
 					class="px-4!"
 					@click="onOnboarding"
 				>
@@ -155,6 +156,47 @@
 			</template>
 		</div>
 	</el-drawer>
+
+	<el-dialog
+		v-model="wizardPluginDialogVisible"
+		:title="t('spacesModule.headings.selectPlugin')"
+		width="520px"
+	>
+		<div class="flex flex-col gap-2">
+			<el-card
+				v-for="item in wizardOptions"
+				:key="item.value"
+				shadow="hover"
+				class="spaces-wizard-plugin-card"
+				:class="{ 'spaces-wizard-plugin-card--disabled': item.disabled }"
+				body-class="p-4!"
+				@click="!item.disabled && onStartOnboarding(item.value)"
+			>
+				<div class="flex items-start gap-3">
+					<div class="spaces-wizard-plugin-card__icon">
+						<icon icon="mdi:puzzle-outline" />
+					</div>
+					<div class="flex-1 min-w-0 spaces-wizard-plugin-card__content">
+						<h3 class="spaces-wizard-plugin-card__title">
+							{{ item.label }}
+						</h3>
+						<p class="spaces-wizard-plugin-card__description">
+							{{ item.description || '&nbsp;' }}
+						</p>
+					</div>
+					<div class="spaces-wizard-plugin-card__chevron">
+						<icon icon="mdi:chevron-right" />
+					</div>
+				</div>
+			</el-card>
+		</div>
+
+		<template #footer>
+			<el-button @click="wizardPluginDialogVisible = false">
+				{{ t('spacesModule.buttons.cancel.title') }}
+			</el-button>
+		</template>
+	</el-dialog>
 </template>
 
 <script setup lang="ts">
@@ -163,7 +205,7 @@ import { useI18n } from 'vue-i18n';
 import { useMeta } from 'vue-meta';
 import { type RouteLocationResolvedGeneric, useRoute, useRouter } from 'vue-router';
 
-import { ElButton, ElDrawer, ElIcon, ElMessageBox } from 'element-plus';
+import { ElButton, ElCard, ElDialog, ElDrawer, ElIcon, ElMessageBox } from 'element-plus';
 
 import { Icon } from '@iconify/vue';
 
@@ -180,7 +222,7 @@ import {
 } from '../../../common';
 import { ListSpaces } from '../components/components';
 import ListSpacesAdjust from '../components/list-spaces-adjust.vue';
-import { useSpacesActions, useSpacesDataSource } from '../composables';
+import { useSpacesActions, useSpacesDataSource, useSpacesPlugins } from '../composables';
 import { RouteNames } from '../spaces.constants';
 import type { ISpace } from '../store/spaces.store.types';
 
@@ -214,9 +256,11 @@ const {
 	resetFilter,
 } = useSpacesDataSource();
 const spacesActions = useSpacesActions();
+const { wizardOptions } = useSpacesPlugins();
 
 const showDrawer = ref<boolean>(false);
 const adjustList = ref<boolean>(false);
+const wizardPluginDialogVisible = ref<boolean>(false);
 
 const remoteFormChanged = ref<boolean>(false);
 
@@ -336,8 +380,24 @@ const onAdjustList = (): void => {
 };
 
 const onOnboarding = (): void => {
+	const enabledWizardOptions = wizardOptions.value.filter((item) => !item.disabled);
+
+	if (enabledWizardOptions.length === 1 && enabledWizardOptions[0]) {
+		onStartOnboarding(enabledWizardOptions[0].value);
+		return;
+	}
+
+	wizardPluginDialogVisible.value = true;
+};
+
+const onStartOnboarding = (type: string): void => {
+	wizardPluginDialogVisible.value = false;
+
 	router.push({
 		name: RouteNames.SPACES_ONBOARDING,
+		params: {
+			type,
+		},
 	});
 };
 
@@ -357,3 +417,75 @@ watch(
 	}
 );
 </script>
+
+<style scoped lang="scss">
+.spaces-wizard-plugin-card {
+	cursor: pointer;
+	transition: all 0.2s ease;
+
+	&:hover {
+		.spaces-wizard-plugin-card__chevron {
+			opacity: 1;
+			transform: translateX(2px);
+		}
+	}
+}
+
+.spaces-wizard-plugin-card--disabled {
+	cursor: not-allowed;
+	opacity: 0.6;
+}
+
+.spaces-wizard-plugin-card__icon {
+	display: flex;
+	align-items: center;
+	justify-content: center;
+	width: 2.5rem;
+	height: 2.5rem;
+	border-radius: var(--el-border-radius-base);
+	background-color: var(--el-color-success-light-9);
+	color: var(--el-color-success);
+	font-size: 1.25rem;
+	flex-shrink: 0;
+}
+
+.spaces-wizard-plugin-card__content {
+	display: flex;
+	flex-direction: column;
+	min-height: 4.5rem;
+}
+
+.spaces-wizard-plugin-card__title {
+	margin: 0 0 0.25rem 0;
+	font-size: 0.9375rem;
+	font-weight: 600;
+	line-height: 1.3;
+	color: var(--el-text-color-primary);
+	display: -webkit-box;
+	-webkit-line-clamp: 2;
+	-webkit-box-orient: vertical;
+	overflow: hidden;
+}
+
+.spaces-wizard-plugin-card__description {
+	margin: 0;
+	font-size: 0.8125rem;
+	line-height: 1.4;
+	color: var(--el-text-color-secondary);
+	display: -webkit-box;
+	-webkit-line-clamp: 2;
+	-webkit-box-orient: vertical;
+	overflow: hidden;
+}
+
+.spaces-wizard-plugin-card__chevron {
+	display: flex;
+	align-items: center;
+	color: var(--el-text-color-placeholder);
+	font-size: 1.25rem;
+	opacity: 0.5;
+	transition: all 0.2s ease;
+	flex-shrink: 0;
+	align-self: center;
+}
+</style>
