+++ apps/admin/src/plugins/spaces-home-control/composables/index.ts
@@ -7,3 +7,4 @@ export * from './useSpaceScenes';
 export * from './useSpaceSensorState';
 export * from './useSpaceSuggestion';
 export * from './useSpaceUndo';
+export * from './useSpacesOnboarding';
